+++ packages/framer-motion/src/three-entry.ts
@@ -6,7 +6,6 @@ export type {
 export { AnimationType } from "./render/utils/types"
 export { animations } from "./motion/features/animations"
 export { MotionContext } from "./context/MotionContext"
-export { checkTargetForNewValues } from "./render/utils/setters"
 export { createBox } from "./projection/geometry/models"
 export { calcLength } from "./projection/geometry/delta-calc"
 export { filterProps } from "./render/dom/utils/filter-props"
